+++ tests/unit/test_router.py
@@ -22,9 +22,11 @@ from gordon import router
 
 
 @pytest.fixture()
-def router_inst(inited_plugins, plugin_kwargs):
+def router_inst(loaded_config, inited_plugins, plugin_kwargs):
     inited_plugins.pop('runnable')
-    router_inst = router.GordonRouter(plugins=inited_plugins, **plugin_kwargs)
+    route_config = loaded_config['core']['route']
+    router_inst = router.GordonRouter(
+            route_config, plugins=inited_plugins.values(), **plugin_kwargs)
     return router_inst
 
 
@@ -54,26 +56,34 @@ def event_msg(mocker):
     return msg
 
 
-@pytest.mark.parametrize('enricher', [True, False])
-def test_init_router(enricher, inited_plugins, plugin_kwargs):
-    inited_plugins.pop('runnable')
-    if not enricher:
-        inited_plugins.pop('enricher')
-
-    router_inst = router.GordonRouter(plugins=inited_plugins, **plugin_kwargs)
-
+@pytest.mark.parametrize('has_enricher', [True, False])
+def test_init_router(has_enricher, inited_plugins, plugin_kwargs):
     exp_phase_route = {
         'consume': 'publish',
         'publish': 'cleanup',
         'cleanup': 'cleanup'
     }
-    if enricher:
+
+    exp_phase_map = {
+        'consume': interfaces.IRunnable,
+        'publish': interfaces.IMessageHandler,
+        'cleanup': interfaces.IMessageHandler
+    }
+
+    inited_plugins.pop('runnable')
+    if has_enricher:
         exp_phase_route['consume'] = 'enrich'
         exp_phase_route['enrich'] = 'publish'
+        exp_phase_map['enrich'] = interfaces.IMessageHandler
+    else:
+        inited_plugins.pop('enricher')
+
+    router_inst = router.GordonRouter(
+        exp_phase_route, plugins=inited_plugins.values(), **plugin_kwargs)
 
     assert exp_phase_route == router_inst.phase_route
-    for _, plugin in router_inst.plugins.items():
-        assert hasattr(plugin, 'handle_message')
+    for phase, plugin in router_inst.phase_plugin_map.items():
+        assert exp_phase_map[phase].providedBy(plugin)
 
 
 @pytest.mark.parametrize('current_phase,exp_next_phase', (
@@ -96,6 +106,7 @@ def test_get_next_phase(current_phase, exp_next_phase, router_inst, event_msg,
 ))
 async def test_route(raises, exp_call_count, exp_log_count, event_msg,
                      router_inst, caplog, monkeypatch):
+
     mock_process_call_count = 0
     mock_process_call_args = []
 
@@ -109,7 +120,7 @@ async def test_route(raises, exp_call_count, exp_log_count, event_msg,
             raise Exception('foo')
 
     monkeypatch.setattr(
-        router_inst.plugins['enricher'], 'handle_message', mock_process)
+        router_inst.phase_plugin_map['enrich'], 'handle_message', mock_process)
 
     await router_inst._route(event_msg)
 
@@ -117,8 +128,18 @@ async def test_route(raises, exp_call_count, exp_log_count, event_msg,
     assert [((event_msg,), {})] == mock_process_call_args
     assert 1 == event_msg._mock_append_to_history
     assert exp_log_count == len(caplog.records)
-    actual_count = router_inst.plugins['event_consumer']._mock_cleanup_count
+    actual_count = \
+        router_inst.phase_plugin_map['cleanup']._mock_handle_message_count
     assert exp_call_count == actual_count
+
+
+@pytest.mark.asyncio
+async def test_route_no_phase_no_cleanup(event_msg, router_inst, caplog):
+    del(router_inst.phase_plugin_map['cleanup'])
+    event_msg.phase = 'blahblah'  # doesnt exist
+
+    await router_inst._route(event_msg)
+    assert 2 == len(caplog.records)  # unknown phase, drop
 
 
 @pytest.mark.asyncio